requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
python-multipart
Pillow
requests
gunicorn